Move ApplicantProgramBlocksController to new `controllers.applicant` package (#265)
* Fix compiler error now that there is a `controllers.applicant` package

* Actually do the refactor properly

* Format
