+++ universal-application-tool-0.0.1/test/controllers/applicant/ApplicantProgramBlocksControllerTest.java
@@ -6,8 +6,6 @@ import static play.mvc.Http.Status.NOT_FOUND;
 import static play.mvc.Http.Status.OK;
 import static play.test.Helpers.fakeRequest;
 
-import controllers.ApplicantProgramBlocksController;
-import controllers.routes;
 import models.Applicant;
 import org.junit.Before;
 import org.junit.Test;
